PythonApi/requirements.txt
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
deep-translator==1.11.4
filelock==3.20.3
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
fsspec==2026.1.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
nltk==3.9.2
numpy==2.4.1
packaging==25.0
pillow==12.1.0
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
soupsieve==2.8.3
tokenizers==0.22.2
tqdm==4.67.1
transformers==4.57.6
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
Flask-SQLAlchemy==3.1.1
psycopg2-binary>=2.9.10
torch
torchvision

pytest==8.4.2
pytest-cov==6.3.0

